ClientGUI listens for SKIN_FILE pref changes
When the preferences change, it clears out the phase display panels, so that
they have to get recreated.  This will update the skins, however it was never
designed to work like this so there are some issues that need addressing.

One is that the AbstractPhaseDisplays will need a die() method, so they can
clear listeners.  Also, there may need to be some cases where things don't
work right.  For instance with MovementDisplay, the beginMyTurn needs to
be called because it normally works on a GameTurnChange event, which won't
happen.

This is the beginning of fixing this, but more work needs to be done.